Update index.qmd
Some proof-reading suggestions to intro. 

One thing to double-check: I took out the "" around "evaluation reports" because these were visible in-text in the report, which I don't think was the intention? But I'm not sure if I might have messed with a formatting code there - best to check! 

One content suggestion: 

Perhaps you could explain the human-in-the-loop models part a bit more? Personally I'm not familiar enough with the topic to really understand what you are talking about there.